Implement client-side templated login screen
Previously we were relying on server-side Heist templating to render
the login screen.  This commit is the first stab at doing this on
client side.  To make it easier, the server side REST protocol
communicates "logged in" information to the client.  Previously REST
handlers simply errored out if not logged in, in the new
implementation the main /rest/app interface will still reply but only
with a response of "not logged in".

This change also removes unnecessary FromJSON instance which were not
used.

This change also gets rid of all Heist templates.