Add 'test' user override for REST API testing
Add ability to run the server in a mode where no login is necessary,
but a user 'test' is force logged-in.

This is for use in REST API unit testing.

When using the 'test' environment, the database is also initialized
only into memory so as to not leave any temporary database files lying
around after tests are done.  A memory database should also be faster
for running tests.