Wrap JSON responses inside a response object
Wrap all JSON responses inside an object that contains login
information and contains the actual response in the o.payload field.

The goal of this change is that each JSON response from the server can
contain information on whether the user is logged in.  This way the
client-side can render the UI for any given object differently
depending on login state.  Any requests to REST entry points that
actually require login (e.g. all create/modify/delete entry points)
will still error out with 403 just like before.

This should make it easier to allow some entry points to be read-only
for public viewing.

Update REST unit tests for this API change.

Also updates the client-side for the API change.

This change also changes request error handling in that we don't pass
around just Strings for error messages but use a new HttpError type
that contains both an HTTP error code and an error message.  This will
be used to signal not only bad requests but authorization errors in
case a user is trying to access another users data.