server: Allow making workouts publicly viewable (#9)
Add a 'public' flag to workouts that signifies its ok to view that
particular instance publicly, e.g. it can be viewed without logging in
or it's viewable by other users.

Adds a new column 'public' to the schema -- this means a DB version
upgrade will happen automatically.

Communicate the owner of 'workout' object to the client-side.  This
way it'll be possible when to render the workout UI as read-only.  The
logic should be that the UI should be read-only when the workout is
not owned by the currently logged in user.

Add a fairly detailed access control tests for the same.